Project implemented as a package to be imported using 'zig fetch'
Alright, I'll just say it so that nobody else does: this is not my
finest work.

However, I've been playing with this for a while now and I'm honestly
kind of tired for now. I'll leave it be for a while and maybe someday I
get back to it.

Major issues I have with it for now:
>  1 - Build system does not make a regular module for import. Instead
it relies on the build system importing. I honestly can't explain it
properly, but check the source code and [this
issue](https://github.com/ziglang/zig/issues/14279). I feel like this is
kind of janky, since my LSP doesn't even detect the import in code, so I
don't get any sort of autocomplete.
>
>  2 - Certain optimize modes throw away important pieces of code, which
is unacceptable. It should optimize all the regular code parts while
keeping the essentials, such as gpio writes and etc.
>
>  3 - Still haven't worked very much on the interrupt overriding
interface. The linker provides the symbols, but the code doesn't really
do anything with it. I want to restructure this module to make these
handlers properly overridable and take advantage of the interrupts array
in the *'supported_chips'* structure.
>
>  4 - The GPIO functions do not work, but using the registers directly
does. In theory they are exactly the same thing, but something is off.

Honestly, the biggest issue for me right now is the lack of LSP support.
The rest is likely easy to fix, but not having an LSP really delays my
coding.

I don't know how or when I'll fix all of these issues or if it's even
possible, but that's for a future moment.