Fixes applied to make the project work with ZLS
This project now is not only a package compatible with the Zig Package
Manager, but also works with the Zig Language Server!

This means that now, when a project that uses UniMicro is created, the
user will have fully functional autocomplete of all the modules
available!

Thanks to [Techatrix](https://github.com/Techatrix) and
[llogick](https://github.com/llogick) for the help, this wouldn't be
possible without them!