Blinky for the stm32f411cc added (#18)
This is the first commit of my blinky example for this chip, it is still
not functional as the build.zig.zon has no dependencies.

It compiles and works when the dependency is properly added, but as of
right now I'm intentionally omitting the UniMicro dependency, since in
the future its URL will be the latest relase URL (but there are still no
releases yet).

NOTE: Since this is the first example ever on this project, there was
still no .gitattributes file as well, so I added it as well.