Improved the interrupts interface (#2)
This project now handles interrupts differently.

Previously, the project created the default interrupt handlers through a
series of PROVIDE directives in the linker script. The handlers were
then overriden by users using an @export() call. However this approach
seemed quite janky, as it relied on the linker script and did stuff out
of zig code that directly affected the compilation process.

Now, the project takes a different approach: the default interrupt
vector is defined as a weak symbol in an object file that's compiled
separately from the rest of the project. This object file is then linked
to the main project again. The user can still override the default
handlers with the interrupts module. This is a more elegant approach,
reliant exclusively on Zig code, without relying on janky and obscure
linker script directives.