Names of the registers and its bits and their documentation are now a perfect match to those of the datasheet
I had preciously changed the names of the registers and its bits a
little, thinking that it would make them more self-explanatory. However,
I've come to the realization that this only made it more difficult to
find them in the datasheet and would confuse users more.

Instead, now all the names are a perfect match to the datasheet and the
documentation for each of these items was straight up copied from there.

This documentation was done with doc-comments, which mean that LSPs can
detect them properly and IDEs can display them. Besides, when the
compiler's documentation generator is finished this will help generating
the library's docs automatically (even though in its current state it
kinda sucks).