Names of registers and bits now match datasheet + HAL functions improved + documentation added
Like with the RCC registers, these were named differently from the
Datasheet, which is now fixed.

Functions for abstracting register access were improved, now there's a
configure pins function that takes one or many pins and its respective
settings, so that in a single function call, all the desired pins from a
Port can be configured.

Beyond that, there's now a function that implements the lock sequence
that prevents the configuration from those pins from being altered.

Documentation was done with Doc-Comments hoping that the compiler's
documentation generator will start to work in the near future. Besides,
this allows LSPs to show this documentation in the editor as well.