Initial verison of set_system_clock function (not working yet)
An auxiliary function was developed to determine the coefficients for
the PLL based on the user's desired system clock. This function seems to
work properly, I tested the coefficients it returned against the
formulas provided by the datasheet and they seem to be correct.

In the case in which this function can't match the desired clock
exactly, it looks for the closest possible result and issues a
compileLog informing that. I'd like for this to be a non-fatal warning,
informing the user but continuing to compile. However, compileLog
statements generate errors and I haven't found a way to do what I want,
so I'll suggest something in the Zig repo and hope it is approved. If
not, I'll revisit this strategy later.

More importantly, the set_system_clock function itself is not working
yet. The coefficients are determined and set through the proper
registers, but when the PLL is actually activated, it just doesn't work.
I'm not sure if it's during the setting of the PLLON bit or the read of
the PLLRDY bit, but somewhere between these two a fault is generated.
I'm also not sure which fault as my debugger is not 100%. I'll look more
into it, but as of right now, this is not yet functional.