Mapped the Flash memory related registers (one of them is probably wrong)
The documentation for these registers is honestly very bad. The worst
offender is the nWRP field in the OPTCR register, which has some of the
poorliest written pieces of text I've ever read in my life.

I'll look more into them eventually and try to make sense of it, but as
of right now it's good enough.

Also I haven't written the HAL functions themselves as I've never really
used this module. When the rest of the funcitonality is done I'll
revisit this.