FIGURED IT OUT!
Finally found out what was causing the problems in tge set_system_clock
function. There's a section of the user manual that talks about the
relationship between the system clock and the flash memory. Apparently,
for each range of values used for the system clock, there's a number of
wait states that need to be configured in the Flash memory.

After this was implemented, the function works flawlessly and the tests
I've done are all successful.

It's worth noting that there's still work to be done in it, since the
aforementioned section of the user manual also relates this number of
wait states to the supply voltage, but as of right now I still don't
know how to determine that. I guess that there's a way to do it using
the Power Controller interface, but I still haven't gotten to that. When
I do, I'll revisit this.

Also worth noting, there was a minor error in the setting of the
peripheral clock divisors, but it has been fixed here as well.